@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.14.0",
-    "@base-sdk/base": "^0.1.0-5",
+    "@base-sdk/base": "^0.1.3",
     "@code-editor/preview-pip": "^0.0.2",
     "@code-editor/vanilla-preview": "^0.0.2",
     "@code-ui/docstring": "^0.0.12",
@@ -19,33 +19,29 @@
     "@emotion/css": "^11.5.0",
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
-    "@frogress/line": "^1.1.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
     "@modulz/design-system": "^0.6.1",
-    "@monaco-editor/react": "^4.1.3",
+    "@monaco-editor/react": "^4.4.1",
+    "@mui/icons-material": "^5.6.1",
+    "@mui/lab": "^5.0.0-alpha.77",
+    "@mui/material": "^5.6.1",
     "@reflect-blocks/figma-embed": "^0.0.5",
-    "@use-gesture/react": "^10.2.2",
+    "@use-gesture/react": "^10.2.11",
     "@visx/gradient": "^1.7.0",
     "@visx/group": "^1.7.0",
     "@visx/hierarchy": "^1.7.0",
     "@visx/shape": "^1.8.0",
-    "codemirror": "^5.61.0",
     "codesandbox": "^2.2.3",
     "cuid": "^2.1.8",
-    "dart-style": "^1.3.2-dev",
     "firebase": "^9.6.0",
     "framer-motion": "^5.3.1",
     "idb": "^6.1.2",
     "moment": "^2.29.1",
-    "monaco-editor": "^0.24.0",
-    "next": "10.0.3",
+    "monaco-editor": "^0.33.0",
+    "next": "^12.1.4",
     "pouchdb-adapter-idb": "^7.2.2",
     "re-resizable": "^6.9.1",
-    "react": "17.0.1",
-    "react-codemirror2": "^7.2.1",
-    "react-dom": "17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-hotkeys-hook": "^3.4.4",
     "react-json-tree": "^0.15.0",
     "react-resizable": "^3.0.1",
@@ -54,7 +50,6 @@
     "recoil": "^0.2.0",
     "rxdb": "^10.5.4",
     "rxjs": "^7.4.0",
-    "styled-components": "^5.3.3",
     "tree-visit": "^0.1.0"
   },
   "devDependencies": {
@@ -66,7 +61,7 @@
     "@zeit/next-css": "^1.0.1",
     "babel-plugin-module-resolver": "^4.1.0",
     "css-loader": "^5.2.4",
-    "next-transpile-modules": "^7.0.0",
+    "next-transpile-modules": "^9.0.0",
     "raw-loader": "^4.0.2",
     "typescript": "^4.2.3"
   }
