@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "@emotion/styled";
-import { ArrowBack } from "@material-ui/icons";
+import { ArrowBack } from "@mui/icons-material";
 import { useRouter } from "next/router";
 import { colors } from "theme";
 import ClientOnly from "components/client-only";
@@ -32,7 +32,6 @@ const RootWrapperAppbarFragmentForSidebar = styled.div`
   flex-direction: row;
   align-items: center;
   gap: 10px;
-  width: 200px;
   align-self: stretch;
   background-color: ${colors.color_editor_bg_on_dark};
   box-sizing: border-box;
