@@ -2,9 +2,9 @@ import React from "react";
 import Link from "next/link";
 import styled from "@emotion/styled";
 
-import FormGroup from "@material-ui/core/FormGroup";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Checkbox from "@material-ui/core/Checkbox";
+import FormGroup from "@mui/material/FormGroup";
+import FormControlLabel from "@mui/material/FormControlLabel";
+import Checkbox from "@mui/material/Checkbox";
 import { WorkspacePreferenceStore } from "store/workspace-preference-store";
 
 export default function PreferencesHomePage() {
