@@ -96,6 +96,9 @@ export function CanvasEventTarget({
   const [first_wheel_event, set_first_wheel_event] =
     useState<FullGestureState<"wheel">>();
 
+  // this is a hack to prevent from onDragStart being called even when no movement is detected.
+  const [drag_start_emitted, set_drag_start_emitted] = useState(false);
+
   useGesture(
     {
       onPinch: onZooming,
@@ -150,7 +153,10 @@ export function CanvasEventTarget({
           return;
         }
 
-        onDragStart(s);
+        if (s.delta[0] || s.delta[1]) {
+          onDragStart(s);
+          set_drag_start_emitted(true);
+        }
       },
       onDrag: (s) => {
         if (isSpacebarPressed) {
@@ -161,6 +167,10 @@ export function CanvasEventTarget({
           return;
         }
 
+        if ((s.delta[0] || s.delta[1]) && !drag_start_emitted) {
+          set_drag_start_emitted(true);
+          onDragStart(s);
+        }
         onDrag(s);
       },
       onDragEnd: (s) => {
@@ -169,8 +179,10 @@ export function CanvasEventTarget({
           return;
         }
 
+        set_drag_start_emitted(false);
         onDragEnd(s);
       },
+      // @ts-ignore
       onMouseDown: onPointerDown,
       onMoveStart: onPointerMoveStart,
       onMoveEnd: onPointerMoveEnd,
@@ -194,8 +206,11 @@ export function CanvasEventTarget({
         overflow: "hidden",
         touchAction: "none",
         cursor: isSpacebarPressed ? "grab" : "default",
+        userSelect: "none",
+        WebkitUserSelect: "none",
       }}
       id="gesture-event-listener"
+      // @ts-ignore
       ref={interactionEventTargetRef}
     >
       {children}
