@@ -24,7 +24,7 @@ export function LazyFrame({
         pointerEvents: "none",
         transition: "opacity 50ms ease-out 0s",
         transformOrigin: "left top",
-        transform: `translateX(${x}px) translateY(${y}px)`,
+        transform: `translate3d(${x}px, ${y}px, 0)`,
         willChange: "transform",
         display: "block",
         position: "fixed",
@@ -40,7 +40,7 @@ export function LazyFrame({
           style={{
             width: "100%",
             height: "100%",
-            display: inViewport ? "block" : "none",
+            contentVisibility: inViewport ? "visible" : "hidden",
           }}
         >
           {children}
