@@ -1,2 +1,4 @@
-export * from "./hovering-target";
+export * from "./target-of-area";
+export * from "./target-of-point";
 export * from "./center-of";
+export * from "./viewbound-edge-scrolling";
