@@ -1,43 +1,116 @@
-import React, { useEffect } from "react";
-import Editor, { useMonaco, Monaco } from "@monaco-editor/react";
+import React, { useRef } from "react";
+import Editor, { OnMount, OnChange } from "@monaco-editor/react";
 import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 import { MonacoEmptyMock } from "./monaco-mock-empty";
+import { register } from "./monaco-utils";
+import { __dangerous__lastFormattedValue__global } from "@code-editor/prettier-services";
+import { debounce } from "utils/debounce";
+import { downloadFile } from "utils/download";
+
+type ICodeEditor = monaco.editor.IStandaloneCodeEditor;
 
 export interface MonacoEditorProps {
-  defaultValue?: string;
-  defaultLanguage?: string;
+  value?: string;
+  language?: string;
+  onChange?: OnChange;
   width?: number | string;
   height?: number | string;
   options?: monaco.editor.IStandaloneEditorConstructionOptions;
 }
 
 export function MonacoEditor(props: MonacoEditorProps) {
-  const monaco: Monaco = useMonaco();
-  useEffect(() => {
-    if (monaco) {
-      setup_react_support(monaco);
-      // monaco.mo
-    }
-  }, [monaco]);
+  const instance = useRef<{ editor: ICodeEditor; format: any } | null>(null);
+
+  const path = "app." + lang2ext(props.language);
+
+  const onMount: OnMount = (editor, monaco) => {
+    const format = editor.getAction("editor.action.formatDocument");
+    const rename = editor.getAction("editor.action.rename");
+
+    instance.current = { editor, format };
+
+    register.initEditor(editor, monaco);
+
+    editor.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.KeyS, function () {
+      format.run();
+    });
+
+    // disabled. todo: find a way to format on new line, but also with adding new line.
+    // editor.addCommand(monaco.KeyCode.Enter, function () {
+    //   // add new line via script, then run format
+    //   format.run();
+    // });
+
+    editor.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.KeyR, function () {
+      // don't reload the entire page, and..
+      // Default is F2
+      rename.run();
+    });
+
+    editor.addAction({
+      // An unique identifier of the contributed action.
+      id: "export-module-as-file",
+
+      // A label of the action that will be presented to the user.
+      label: "Export as file",
+      precondition: null,
+      keybindingContext: null,
+      contextMenuGroupId: "navigation",
+      contextMenuOrder: 1.5,
+      run: function (ed) {
+        downloadFile({ data: ed.getModel().getValue(), filename: path });
+      },
+    });
+
+    editor.onDidChangeModelContent(() =>
+      debounce(() => editor.saveViewState(), 200)
+    );
+  };
 
   return (
     <Editor
+      beforeMount={register.initMonaco}
+      onMount={onMount}
       width={props.width}
       height={props.height}
-      defaultLanguage={
-        pollyfill_language(props.defaultLanguage) ?? "typescript"
-      }
+      language={pollyfill_language(props.language) ?? "typescript"}
+      path={path}
       loading={<MonacoEmptyMock l={5} />}
-      defaultValue={props.defaultValue ?? "// no content"}
+      value={props.value ?? "// no content"}
       theme="vs-dark"
+      onChange={(...v) => {
+        if (v[0] === __dangerous__lastFormattedValue__global) {
+          // if change is caused by formatter, ignore.
+          return;
+        }
+        props.onChange(...v);
+      }}
       options={{
         ...props.options,
         // overrided default options
+        wordWrap: "off",
         unusualLineTerminators: "off",
       }}
     />
   );
 }
+
+const lang2ext = (lang: string) => {
+  switch (lang) {
+    case "typescript":
+      return "ts";
+    case "javascript":
+      return "js";
+    case "tsx":
+      return "tsx";
+    case "jsx":
+      return "jsx";
+    case "dart":
+      return "dart";
+    default:
+      return lang;
+  }
+};
 
 const pollyfill_language = (lang: string) => {
   switch (lang) {
@@ -49,25 +122,5 @@ const pollyfill_language = (lang: string) => {
       return lang;
   }
 };
-
-function setup_react_support(monaco: Monaco) {
-  // adding jsx support - https://github.com/microsoft/monaco-editor/issues/264
-  monaco.languages.typescript.typescriptDefaults.setCompilerOptions({
-    target: monaco.languages.typescript.ScriptTarget.Latest,
-    allowNonTsExtensions: true,
-    moduleResolution: monaco.languages.typescript.ModuleResolutionKind.NodeJs,
-    module: monaco.languages.typescript.ModuleKind.CommonJS,
-    noEmit: true,
-    esModuleInterop: true,
-    jsx: monaco.languages.typescript.JsxEmit.React,
-    reactNamespace: "React",
-    allowJs: true,
-  });
-
-  monaco.languages.typescript.typescriptDefaults.setDiagnosticsOptions({
-    noSemanticValidation: false,
-    noSyntaxValidation: false,
-  });
-}
 
 export { useMonaco } from "@monaco-editor/react";
