@@ -2,5 +2,5 @@ export * as figmacomp from "./figma";
 export * as canvas from "./design-preview-as-is";
 export * as code from "./code-editor";
 export * as runner from "./app-runner";
-export * as visualization from "../../editor-packages/editor-debugger/components/visualization";
+export * as visualization from "../../editor-packages/editor-devtools/components/visualization";
 export * from "./client-only";
