@@ -1,3 +1,6 @@
+import type { FrameworkConfig } from "@designto/config";
+import type { ConsoleLog, EditorState, ScenePreviewData } from "core/states";
+
 export type WorkspaceAction =
   //
   | HistoryAction
@@ -12,19 +15,20 @@ export type HistoryAction =
   | Action;
 
 export type Action =
-  //
   | PageAction
-  //
   | SelectNodeAction
-  //
-  | HighlightLayerAction;
+  | HighlightLayerAction
+  | CanvasModeAction
+  | PreviewAction
+  | CodeEditorAction
+  | DevtoolsAction;
 
 export type ActionType = Action["type"];
 
 export type HierarchyAction = SelectNodeAction;
 export interface SelectNodeAction {
   type: "select-node";
-  node: string;
+  node: string | string[];
 }
 
 export type PageAction = SelectPageAction;
@@ -37,4 +41,47 @@ export interface SelectPageAction {
 export interface HighlightLayerAction {
   type: "highlight-layer";
   id: string;
+}
+
+type CanvasModeAction = CanvasModeSwitchAction | CanvasModeGobackAction;
+export interface CanvasModeSwitchAction {
+  type: "canvas-mode-switch";
+  mode: EditorState["canvasMode"];
+}
+
+export interface CanvasModeGobackAction {
+  type: "canvas-mode-goback";
+  fallback?: EditorState["canvasMode"];
+}
+
+export type PreviewAction = PreviewSetAction | PreviewBuildingStateUpdateAction;
+
+export interface PreviewSetAction {
+  type: "preview-set";
+  data: ScenePreviewData;
+}
+
+export interface PreviewBuildingStateUpdateAction {
+  type: "preview-update-building-state";
+  isBuilding: boolean;
+}
+
+export type CodeEditorAction = CodeEditorEditComponentCodeAction;
+
+export interface CodeEditorEditComponentCodeAction {
+  type: "code-editor-edit-component-code";
+  id: string;
+  framework: FrameworkConfig["framework"];
+  componentName: string;
+  raw: string;
+}
+
+export type DevtoolsAction = DevtoolsConsoleAction | DevtoolsConsoleClearAction;
+export interface DevtoolsConsoleAction {
+  type: "devtools-console";
+  log: ConsoleLog;
+}
+
+export interface DevtoolsConsoleClearAction {
+  type: "devtools-console-clear";
 }
