@@ -1,23 +1,45 @@
 import produce from "immer";
-import { Action, SelectNodeAction, SelectPageAction } from "core/actions";
+import type {
+  Action,
+  SelectNodeAction,
+  SelectPageAction,
+  CodeEditorEditComponentCodeAction,
+  CanvasModeSwitchAction,
+  CanvasModeGobackAction,
+  PreviewBuildingStateUpdateAction,
+  PreviewSetAction,
+  DevtoolsConsoleAction,
+  DevtoolsConsoleClearAction,
+} from "core/actions";
 import { EditorState } from "core/states";
 import { useRouter } from "next/router";
 import { CanvasStateStore } from "@code-editor/canvas/stores";
+import assert from "assert";
+
+const _editor_path_name = "/files/[key]/";
 
 export function editorReducer(state: EditorState, action: Action): EditorState {
   const router = useRouter();
   const filekey = state.design.key;
 
-  // TODO: handle actions here.
   switch (action.type) {
     case "select-node": {
       const { node } = <SelectNodeAction>action;
       console.clear();
       console.info("cleard console by editorReducer#select-node");
 
+      const ids = Array.isArray(node) ? node : [node];
+      const primary = ids?.[0];
+
       // update router
-      router.query.node = node ?? state.selectedPage;
-      router.push(router);
+      router.push(
+        {
+          pathname: _editor_path_name,
+          query: { ...router.query, node: primary ?? state.selectedPage },
+        },
+        undefined,
+        { shallow: true }
+      );
 
       return produce(state, (draft) => {
         const _canvas_state_store = new CanvasStateStore(
@@ -25,7 +47,7 @@ export function editorReducer(state: EditorState, action: Action): EditorState {
           state.selectedPage
         );
 
-        const new_selections = [node].filter(Boolean);
+        const new_selections = ids.filter(Boolean);
         _canvas_state_store.saveLastSelection(...new_selections);
 
         // assign new nodes set to the state.
@@ -42,8 +64,14 @@ export function editorReducer(state: EditorState, action: Action): EditorState {
       console.info("cleard console by editorReducer#select-page");
 
       // update router
-      router.query.node = page;
-      router.push(router);
+      router.push(
+        {
+          pathname: _editor_path_name,
+          query: { ...router.query, node: page },
+        },
+        undefined,
+        { shallow: true }
+      );
 
       return produce(state, (draft) => {
         const _canvas_state_store = new CanvasStateStore(filekey, page);
@@ -58,9 +86,109 @@ export function editorReducer(state: EditorState, action: Action): EditorState {
         draft.selectedNodes = last_known_selections_of_this_page;
       });
     }
+    case "code-editor-edit-component-code": {
+      const { ...rest } = <CodeEditorEditComponentCodeAction>action;
+      return produce(state, (draft) => {
+        draft.editingModule = {
+          ...rest,
+          type: "single-file-component",
+          lang: "unknown",
+        };
+      });
+      //
+    }
+    case "canvas-mode-switch": {
+      const { mode } = <CanvasModeSwitchAction>action;
+
+      router.push({
+        pathname: _editor_path_name,
+        query: { ...router.query, mode: mode },
+        // no need to set shallow here.
+      });
+
+      return produce(state, (draft) => {
+        draft.canvasMode_previous = draft.canvasMode;
+        draft.canvasMode = mode;
+      });
+    }
+    case "canvas-mode-goback": {
+      const { fallback } = <CanvasModeGobackAction>action;
+
+      const dest = state.canvasMode_previous ?? fallback;
+
+      router.push({
+        pathname: _editor_path_name,
+        query: { ...router.query, mode: dest },
+        // no need to set shallow here.
+      });
+
+      return produce(state, (draft) => {
+        assert(
+          dest,
+          "canvas-mode-goback: cannot resolve destination. (no fallback provided)"
+        );
+        draft.canvasMode_previous = draft.canvasMode; // swap
+        draft.canvasMode = dest; // previous or fallback
+      });
+    }
+    case "preview-update-building-state": {
+      const { isBuilding } = <PreviewBuildingStateUpdateAction>action;
+      return produce(state, (draft) => {
+        if (draft.currentPreview) {
+          draft.currentPreview.isBuilding = isBuilding;
+        } else {
+          draft.currentPreview = {
+            loader: null,
+            viewtype: "unknown",
+            isBuilding: true,
+            widgetKey: null,
+            componentName: null,
+            fallbackSource: "loading",
+            initialSize: null,
+            meta: null,
+            source: null,
+            updatedAt: Date.now(),
+          };
+        }
+      });
+    }
+    case "preview-set": {
+      const { data } = <PreviewSetAction>action;
+      return produce(state, (draft) => {
+        draft.currentPreview = data; // set
+      });
+    }
+    case "devtools-console": {
+      const { log } = <DevtoolsConsoleAction>action;
+      return produce(state, (draft) => {
+        if (!draft.devtoolsConsole?.logs?.length) {
+          draft.devtoolsConsole = { logs: [] };
+        }
+
+        const logs = Array.from(state.devtoolsConsole?.logs ?? []);
+        logs.push(log);
+
+        draft.devtoolsConsole.logs = logs;
+      });
+      break;
+    }
+    case "devtools-console-clear": {
+      const {} = <DevtoolsConsoleClearAction>action;
+      return produce(state, (draft) => {
+        if (draft.devtoolsConsole?.logs?.length) {
+          draft.devtoolsConsole.logs = [
+            {
+              id: "clear",
+              method: "info",
+              data: ["Console was cleared"],
+            },
+          ];
+        }
+      });
+      break;
+    }
     default:
       throw new Error(`Unhandled action type: ${action["type"]}`);
   }
-
   return state;
 }
