@@ -14,7 +14,6 @@ import { RemoteDesignSessionCacheStore } from "../store";
 import { convert } from "@design-sdk/figma-node-conversion";
 import { mapFigmaRemoteToFigma } from "@design-sdk/figma-remote/lib/mapper";
 import { useFigmaAccessToken } from ".";
-import { FileResponse } from "@design-sdk/figma-remote-types";
 import {
   FigmaDesignRepository,
   TFetchFileForApp,
@@ -102,8 +101,8 @@ export function useDesign({
     }
 
     if (targetnodeconfig) {
+      const filekey = targetnodeconfig.file;
       // load design from local storage or remote figma
-
       const cacheStore = new RemoteDesignSessionCacheStore({
         file: targetnodeconfig.file,
         node: targetnodeconfig.node,
@@ -115,7 +114,10 @@ export function useDesign({
           last_response.nodes[targetnodeconfig.node]
         );
         const _2_converted_to_reflect = convert.intoReflectNode(
-          _1_converted_to_figma
+          _1_converted_to_figma,
+          null,
+          "rest",
+          filekey
         );
 
         const res = <TargetNodeConfig>{
@@ -172,7 +174,7 @@ export function useDesign({
   return design;
 }
 
-type TUseDesignFile =
+export type TUseDesignFile =
   | TFetchFileForApp
   | {
       __type: "error";
