@@ -1,6 +1,6 @@
-import styled from "@emotion/styled";
-import React from "react";
-
+import { css, jsx } from "@emotion/react";
+import React, { CSSProperties } from "react";
+import { Resizable, ResizableProps } from "re-resizable";
 /**
  * Wrapper for WorkspaceContentPanelGridLayout
  * @param props
@@ -11,42 +11,52 @@ export function WorkspaceContentPanel({
   disableBorder = true,
   flex = 1,
   zIndex,
+  overflow = "auto",
   backgroundColor = "none",
   hidden = false,
+  resize = {
+    top: false,
+    right: false,
+    bottom: false,
+    left: false,
+  },
+  minWidth = 0,
 }: {
-  backgroundColor?: string;
   children: JSX.Element;
-  disableBorder?: boolean;
-  flex?: number;
-  zIndex?: number;
   hidden?: boolean;
-}) {
+} & Omit<WorkspaceCPanelStyleProps, "display"> & {
+    resize?: ResizableProps["enable"];
+    minWidth?: number;
+  }) {
+  const [oflex, setOFlex] = React.useState(flex);
+
   return (
-    <WorkspaceCPanel
-      display={hidden ? "none" : undefined}
-      flex={flex}
-      zIndex={zIndex}
-      backgroundColor={backgroundColor}
-      disableBorder={disableBorder}
+    <Resizable
+      enable={resize}
+      onResizeStart={() => setOFlex(undefined)}
+      minWidth={minWidth}
+      style={{
+        minWidth: minWidth,
+        border: disableBorder ? "none" : "solid #d2d2d2",
+        backgroundColor: backgroundColor,
+        borderWidth: 1,
+        alignSelf: "stretch",
+        flex: oflex,
+        overflow: overflow,
+        display: hidden ? "none" : undefined,
+        zIndex: zIndex ?? 0,
+      }}
     >
       {children}
-    </WorkspaceCPanel>
+    </Resizable>
   );
 }
 
-const WorkspaceCPanel = styled.div<{
+type WorkspaceCPanelStyleProps = {
   flex?: number;
-  backgroundColor: string;
-  disableBorder: boolean;
+  overflow?: CSSProperties["overflow"];
+  backgroundColor?: string;
+  disableBorder?: boolean;
   zIndex?: number;
   display: "none" | undefined;
-}>`
-  border: ${(p) => (p.disableBorder ? "none" : "solid #d2d2d2")};
-  background-color: ${(p) => p.backgroundColor};
-  border-width: 1px;
-  align-self: stretch;
-  flex: ${(p) => p.flex};
-  overflow: auto;
-  ${(p) => p.display === "none" && "display: none;"}
-  z-index: ${(p) => p.zIndex ?? 0};
-`;
+};
