@@ -1,6 +1,5 @@
 /// <reference types="next" />
-/// <reference types="next/types/global" />
-declare module "*.txt" {
-  const content: string;
-  export default content;
-}
+/// <reference types="next/image-types/global" />
+
+// NOTE: This file should not be edited
+// see https://nextjs.org/docs/basic-features/typescript for more information.
