@@ -1,9 +1,11 @@
 const TerserPlugin = require("terser-webpack-plugin");
 const withTM = require("next-transpile-modules")([
+  "dirty",
+
   // region @editor-app
   "@editor-app/live-session",
   "@code-editor/preview-pip", // TODO: remove me. this is for development. for production, use npm ver instead.
-  "@code-editor/debugger",
+  "@code-editor/devtools",
   "@code-editor/canvas",
 
   // region editor-submodule deps
@@ -12,6 +14,7 @@ const withTM = require("next-transpile-modules")([
 
   // -----------------------------
   // region @designto-code
+  "@designto/debugger",
   "@designto/config",
   "@grida/builder-config-preset",
   "@grida/builder-platform-types",
@@ -19,12 +22,15 @@ const withTM = require("next-transpile-modules")([
   "@designto/sanitized",
   "@designto/token",
   "@designto/flutter",
+  "@designto/solid-js",
   "@designto/web",
   "@designto/vanilla",
   "@designto/react",
   "@designto/react-native",
 
   "@code-features/assets",
+  "@code-features/module",
+  "@code-features/documentation",
   "@code-features/component",
   "@code-features/flags",
   // -----------------------------
@@ -85,6 +91,9 @@ const withTM = require("next-transpile-modules")([
   // region web builders
   "@web-builder/nodejs",
   "@web-builder/core",
+  "@web-builder/module-es",
+  "@web-builder/module-jsx",
+  "@web-builder/solid-js",
   "@web-builder/vanilla",
   "@web-builder/react-core",
   "@web-builder/react",
@@ -104,11 +113,25 @@ module.exports = withTM({
       include: /node_modules/,
     });
 
-    const terser = config.optimization.minimizer.find(
-      (m) => m.constructor.name === "TerserPlugin"
-    );
+    config.resolve.fallback = {
+      fs: false, // used by handlebars
+      path: false, // used by handlebars
+      crypto: false, // or crypto-browserify (used for totp auth)
+      stream: false, // or stream-browserify (used for totp auth)
+    };
+
+    // -----------------------------
     // for @flutter-builder classname issue
-    terser.options.terserOptions.keep_classnames = true;
+    config.optimization.minimizer.push(
+      new TerserPlugin({
+        parallel: true,
+        terserOptions: {
+          // https://github.com/webpack-contrib/terser-webpack-plugin#terseroptions
+          keep_classnames: true,
+        },
+      })
+    );
+    // -----------------------------
 
     return config;
   },
