@@ -1,20 +1,59 @@
 import React, { useEffect, useCallback, useReducer, useState } from "react";
-import { useRouter } from "next/router";
+import { useRouter, NextRouter } from "next/router";
 import { SigninToContinueBannerPrmoptProvider } from "components/prompt-banner-signin-to-continue";
 import { Editor, EditorDefaultProviders } from "scaffolds/editor";
 import { EditorSnapshot, StateProvider } from "core/states";
 import { WorkspaceAction } from "core/actions";
-import { useDesignFile } from "hooks";
-
+import { useDesignFile, TUseDesignFile } from "hooks";
 import { warmup } from "scaffolds/editor";
-import { FileResponse } from "@design-sdk/figma-remote-types";
+import { EditorBrowserMetaHead } from "components/editor";
+import type { FileResponse } from "@design-sdk/figma-remote-types";
 
 export default function FileEntryEditor() {
   const router = useRouter();
-  const { key, node } = router.query;
-  const filekey = key as string;
-  const nodeid = node as string;
+  const { key } = router.query;
 
+  const [nodeid, setNodeid] = useState<string>();
+  const filekey = key as string;
+  // const nodeid = node as string;
+
+  // background whole file fetching
+  const file = useDesignFile({ file: filekey });
+
+  useEffect(() => {
+    if (!router.isReady) return;
+
+    if (!nodeid) {
+      // set nodeid only first time
+      setNodeid(router.query.node as string);
+      console.log("nodeid set", router.query.node);
+    }
+  }, [router.isReady]);
+
+  return (
+    <SigninToContinueBannerPrmoptProvider>
+      <SetupEditor
+        key={filekey}
+        file={file}
+        filekey={filekey}
+        nodeid={nodeid}
+        router={router}
+      />
+    </SigninToContinueBannerPrmoptProvider>
+  );
+}
+
+function SetupEditor({
+  filekey,
+  nodeid,
+  router,
+  file,
+}: {
+  nodeid: string;
+  filekey: string;
+  router: NextRouter;
+  file: TUseDesignFile;
+}) {
   const [loading, setLoading] = useState<boolean>(true);
 
   const [initialState, initialDispatcher] = useReducer(warmup.initialReducer, {
@@ -25,20 +64,21 @@ export default function FileEntryEditor() {
     initialDispatcher({ type: "update", value: action });
   }, []);
 
-  // background whole file fetching
-  const file = useDesignFile({ file: filekey });
-
-  const prevstate =
-    initialState.type == "success" && initialState.value.history.present;
+  const initialCanvasMode = q_map_canvas_mode_from_query(
+    router.query.mode as string
+  ); // TODO: change this to reflect the nodeid input
 
   const initWith = (file: FileResponse) => {
+    const prevstate =
+      initialState.type == "success" && initialState.value.history.present;
+
     let val: EditorSnapshot;
 
     // TODO: seed this as well
     // ->> file.styles;
 
     const components = warmup.componentsFrom(file);
-    const pages = warmup.pagesFrom(file);
+    const pages = warmup.pagesFrom(filekey, file);
 
     if (prevstate) {
       val = {
@@ -64,12 +104,14 @@ export default function FileEntryEditor() {
         selectedPage: warmup.selectedPage(prevstate, pages, nodeid && [nodeid]),
         selectedLayersOnPreview: [],
         design: {
+          name: file.name,
           input: null,
           components: components,
           // styles: null,
           key: filekey,
           pages: pages,
         },
+        canvasMode: initialCanvasMode,
       };
     }
 
@@ -127,14 +169,15 @@ export default function FileEntryEditor() {
   ]);
 
   const safe_value = warmup.safestate(initialState);
+
   return (
-    <SigninToContinueBannerPrmoptProvider>
-      <StateProvider state={safe_value} dispatch={handleDispatch}>
-        <EditorDefaultProviders>
+    <StateProvider state={safe_value} dispatch={handleDispatch}>
+      <EditorDefaultProviders>
+        <EditorBrowserMetaHead>
           <Editor loading={loading} />
-        </EditorDefaultProviders>
-      </StateProvider>
-    </SigninToContinueBannerPrmoptProvider>
+        </EditorBrowserMetaHead>
+      </EditorDefaultProviders>
+    </StateProvider>
   );
 }
 
@@ -181,3 +224,25 @@ export default function FileEntryEditor() {
     
   }
  */
+
+const q_map_canvas_mode_from_query = (
+  mode: string
+): EditorSnapshot["canvasMode"] => {
+  switch (mode) {
+    case "free":
+    case "isolated-view":
+    case "fullscreen-preview":
+      return mode;
+
+    // -------------------------
+    // legacy query param key
+    case "full":
+      return "free";
+    case "isolate":
+      return "isolated-view";
+    // -------------------------
+
+    default:
+      return "free";
+  }
+};
