@@ -1,13 +1,22 @@
+import React from "react";
+import Head from "next/head";
+
 import { HomeInput } from "scaffolds/home-input";
 import { HomeDashboard } from "scaffolds/home-dashboard";
-import React from "react";
 import { useAuthState } from "hooks/use-auth-state";
 
 export default function Home() {
   const authstate = useAuthState();
 
   // region - dev injected
-  return <HomeDashboard />;
+  return (
+    <>
+      <Head>
+        <title>Grida | Home</title>
+      </Head>
+      <HomeDashboard />
+    </>
+  );
   // endregion
 
   switch (authstate) {
