@@ -1,82 +1,106 @@
-import React, { useEffect, useState } from "react";
+import React, { useCallback } from "react";
 import styled from "@emotion/styled";
 import { Canvas } from "@code-editor/canvas";
 import { useEditorState, useWorkspace } from "core/states";
-import { Preview } from "scaffolds/preview";
+import {
+  D2CVanillaPreview,
+  WebWorkerD2CVanillaPreview,
+} from "scaffolds/preview";
 import useMeasure from "react-use-measure";
 import { useDispatch } from "core/dispatch";
 import { FrameTitleRenderer } from "./render/frame-title";
 import { IsolateModeCanvas } from "./isolate-mode";
-import { useRouter } from "next/router";
-
-type ViewMode = "full" | "isolate";
+import { Dialog } from "@mui/material";
+import { FullScreenPreview } from "scaffolds/preview-full-screen";
 
 /**
  * Statefull canvas segment that contains canvas as a child, with state-data connected.
  */
 export function VisualContentArea() {
   const [state] = useEditorState();
-  const router = useRouter();
-  const { mode: q_mode } = router.query;
-
-  // this hook is used for focusing the node on the first load with the initial selection is provided externally.
-  useEffect(() => {
-    // if the initial selection is available, and not empty &&
-    if (state.selectedNodesInitial?.length && q_mode == "isolate") {
-      // trigger isolation mode once.
-      setMode("isolate");
-
-      // TODO: set explicit canvas initial transform.
-      // make the canvas fit to the initial target even when the isolation mode is complete by the user.
-    }
-  }, [state.selectedNodesInitial]);
-
   const [canvasSizingRef, canvasBounds] = useMeasure();
 
   const { highlightedLayer, highlightLayer } = useWorkspace();
   const dispatch = useDispatch();
 
-  const { selectedPage, design, selectedNodes } = state;
+  const {
+    selectedPage,
+    design,
+    selectedNodes,
+    canvasMode,
+    canvasMode_previous,
+  } = state;
 
   const thisPageNodes = selectedPage
-    ? state.design.pages
-        .find((p) => p.id == selectedPage)
-        .children.filter(Boolean)
+    ? design.pages.find((p) => p.id == selectedPage).children.filter(Boolean)
     : [];
 
   const isEmptyPage = thisPageNodes?.length === 0;
 
-  const [mode, _setMode] = useState<ViewMode>("full");
+  const startIsolatedViewMode = useCallback(
+    () =>
+      dispatch({
+        type: "canvas-mode-switch",
+        mode: "isolated-view",
+      }),
+    [dispatch]
+  );
 
-  const setMode = (m: ViewMode) => {
-    _setMode(m);
+  const startFullscreenPreviewMode = useCallback(
+    () =>
+      dispatch({
+        type: "canvas-mode-switch",
+        mode: "fullscreen-preview",
+      }),
+    [dispatch]
+  );
 
-    // update the router
-    (router.query.mode = m) && router.push(router);
-  };
+  const endIsolatedViewMode = useCallback(
+    () =>
+      dispatch({
+        type: "canvas-mode-switch",
+        mode: "free",
+      }),
+    [dispatch]
+  );
+
+  const exitFullscreenPreview = useCallback(
+    () =>
+      dispatch({
+        type: "canvas-mode-goback",
+        fallback: "isolated-view",
+      }),
+    [dispatch]
+  );
 
   return (
-    <CanvasContainer
-      ref={canvasSizingRef}
-      id="canvas"
-      maxWidth={mode == "isolate" ? "calc((100vw - 200px) * 0.6)" : "100%"} // TODO: make this dynamic
-    >
+    <CanvasContainer ref={canvasSizingRef} id="canvas">
       {/* <EditorAppbarFragments.Canvas /> */}
 
       {isEmptyPage ? (
         <></>
       ) : (
         <>
-          {mode == "isolate" && (
+          <FullScreenPreviewContainer
+            show={canvasMode == "fullscreen-preview"}
+            onExit={exitFullscreenPreview}
+          />
+          {(canvasMode == "isolated-view" ||
+            (canvasMode == "fullscreen-preview" &&
+              canvasMode_previous === "isolated-view")) && (
             <IsolateModeCanvas
-              onClose={() => {
-                setMode("full");
-              }}
+              hidden={
+                // if prev mode is this, hide, not remove.
+                canvasMode == "fullscreen-preview" &&
+                canvasMode_previous === "isolated-view"
+              }
+              onClose={endIsolatedViewMode}
+              onEnterFullscreen={startFullscreenPreviewMode}
             />
           )}
           <div
             style={{
-              display: mode == "full" ? undefined : "none",
+              display: canvasMode !== "free" && "none",
             }}
           >
             <Canvas
@@ -89,10 +113,10 @@ export function VisualContentArea() {
               ]}
               filekey={state.design.key}
               pageid={selectedPage}
-              selectedNodes={selectedNodes.filter(Boolean)}
+              selectedNodes={selectedNodes}
               highlightedLayer={highlightedLayer}
-              onSelectNode={(node) => {
-                dispatch({ type: "select-node", node: node?.id });
+              onSelectNode={(...nodes) => {
+                dispatch({ type: "select-node", node: nodes.map((n) => n.id) });
               }}
               onClearSelection={() => {
                 dispatch({ type: "select-node", node: null });
@@ -100,21 +124,27 @@ export function VisualContentArea() {
               nodes={thisPageNodes}
               // initialTransform={ } // TODO: if the initial selection is provided from first load, from the query param, we have to focus to fit that node.
               renderItem={(p) => {
-                return <Preview key={p.node.id} target={p.node} {...p} />;
+                return (
+                  // <WebWorkerD2CVanillaPreview
+                  //   key={p.node.id}
+                  //   target={p.node}
+                  //   {...p}
+                  // />
+                  <D2CVanillaPreview key={p.node.id} target={p.node} {...p} />
+                );
               }}
               config={{
                 can_highlight_selected_layer: true,
                 marquee: {
-                  disabled: true,
+                  disabled: false,
                 },
               }}
               renderFrameTitle={(p) => (
                 <FrameTitleRenderer
                   key={p.id}
                   {...p}
-                  onRunClick={() => {
-                    setMode("isolate");
-                  }}
+                  runnable={selectedNodes.length === 1}
+                  onRunClick={startIsolatedViewMode}
                 />
               )}
             />
@@ -125,11 +155,22 @@ export function VisualContentArea() {
   );
 }
 
-const CanvasContainer = styled.div<{
-  maxWidth?: string;
-}>`
+function FullScreenPreviewContainer({
+  onExit,
+  show,
+}: {
+  onExit: () => void;
+  show: boolean;
+}) {
+  return (
+    <Dialog fullScreen onClose={onExit} open={show}>
+      <FullScreenPreview onClose={onExit} />
+    </Dialog>
+  );
+}
+
+const CanvasContainer = styled.div`
   display: flex;
   flex-direction: column;
-  max-width: ${(p) => p.maxWidth};
   height: 100%;
 `;
