@@ -12,15 +12,18 @@ import {
   MainImageRepository,
 } from "@design-sdk/core/assets-repository";
 import { useEditorState, useWorkspaceState } from "core/states";
-import { utils_dart } from "utils";
+import { useDispatch } from "core/dispatch";
 import type { ReflectSceneNode } from "@design-sdk/core";
 import { RemoteImageRepositories } from "@design-sdk/figma-remote/lib/asset-repository/image-repository";
 import { useTargetContainer } from "hooks/use-target-node";
 import assert from "assert";
+import { debounce } from "utils/debounce";
+import { supportsScripting } from "config";
 
 export function CodeSegment() {
   const router = useRouter();
   const [result, setResult] = useState<Result>();
+  const dispatch = useDispatch();
   const wstate = useWorkspaceState();
   const [state] = useEditorState();
   const [framework_config, set_framework_config] = useState(
@@ -43,15 +46,9 @@ export function CodeSegment() {
     if (
       result.framework.framework !==
         targetStateRef?.current?.config.framework ||
-      result.id !== targetStateRef?.current?.node.id
+      result.id !== targetStateRef?.current?.node?.id
     ) {
       return;
-    }
-
-    if (framework_config.language == "dart") {
-      // special formatter support for dartlang
-      result.code.raw = utils_dart.format(result.code.raw);
-      result.scaffold.raw = utils_dart.format(result.scaffold.raw);
     }
 
     setResult(result);
@@ -114,6 +111,26 @@ export function CodeSegment() {
     }
   }, [targetted?.id, framework_config]);
 
+  const onChangeHandler = debounce((k, code) => {
+    if (!result) {
+      return;
+    }
+    if (!targetted) {
+      return;
+    }
+
+    // currently react and vanilla are supported
+    if (supportsScripting(framework_config.framework)) {
+      dispatch({
+        type: "code-editor-edit-component-code",
+        framework: framework_config.framework,
+        componentName: result.name,
+        id: targetted.id,
+        raw: code,
+      });
+    }
+  }, 500);
+
   const { code, scaffold, name: componentName } = result ?? {};
   return (
     <CodeEditorContainer>
@@ -157,6 +174,18 @@ export function CodeSegment() {
               }
               break;
             }
+            case "solid-js": {
+              switch (o.styling) {
+                case "styled-components":
+                  c = get_framework_config("solid-with-styled-components");
+                  break;
+                case "inline-css": {
+                  c = get_framework_config("solid-with-inline-css");
+                  break;
+                }
+              }
+              break;
+            }
             case "flutter":
               c = get_framework_config(o.framework);
               break;
@@ -175,13 +204,15 @@ export function CodeSegment() {
         options={{
           automaticLayout: true,
         }}
+        onChange={onChangeHandler}
         files={
           code
             ? {
-                "index.tsx": {
+                // TODO: make this to match framework
+                "App.tsx": {
                   raw: code.raw,
                   language: framework_config.language,
-                  name: "index.tsx",
+                  name: "App.tsx",
                 },
               }
             : {
